smartfilevault/requirements.txt
flask==2.3.3
requests==2.31.0
pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-docx==1.0.1
pypdf2==3.0.1
nltk==3.8.1